#43 🐛Fix: styled-component S-Dot 네이밍으로 변경
