+++ src/pages/HomePage.tsx
@@ -1,7 +1,4 @@
-import React from 'react';
-import NavBar from '@/components/layouts/PageNavBar';
-
-const HomePage: React.FC = () => {
+const HomePage = () => {
   return (
     <div>
       <div>This Route is HomePage</div>
